+++ Source/Library/Collection/HCIsCollectionOnlyContaining.h
@@ -7,10 +7,10 @@
 //  Source: https://github.com/hamcrest/OCHamcrest
 //
 
-#import <OCHamcrest/HCBaseMatcher.h>
+#import <OCHamcrest/HCDiagnosingMatcher.h>
 
 
-@interface HCIsCollectionOnlyContaining : HCBaseMatcher
+@interface HCIsCollectionOnlyContaining : HCDiagnosingMatcher
 
 + (instancetype)isCollectionOnlyContaining:(id <HCMatcher>)matcher;
 - (instancetype)initWithMatcher:(id <HCMatcher>)matcher;
@@ -23,25 +23,25 @@ FOUNDATION_EXPORT id HC_onlyContains(id itemMatch, ...) NS_REQUIRES_NIL_TERMINAT
 /**
  onlyContains(firstMatcher, ...) -
  Matches if each element of collection satisfies any of the given matchers.
- 
+
  @param firstMatcher,...  A comma-separated list of matchers ending with @c nil.
- 
+
  This matcher iterates the evaluated collection, confirming whether each element satisfies any of
  the given matchers.
- 
+
  Any argument that is not a matcher is implicitly wrapped in an @ref equalTo matcher to check for
  equality.
- 
+
  Example:
- 
+
  @par
  @ref onlyContains(startsWith(@"Jo"), nil)
- 
+
  will match a collection [@"Jon", @"John", @"Johann"].
- 
+
  (In the event of a name clash, don't \#define @c HC_SHORTHAND and use the synonym
  @c HC_onlyContains instead.)
- 
+
  @ingroup collection_matchers
  */
 #ifdef HC_SHORTHAND