+++ Source/Library/Collection/HCIsCollectionOnlyContaining.m
@@ -32,18 +32,46 @@
     return self;
 }
 
-- (BOOL)matches:(id)collection
+- (BOOL)matches:(id)collection describingMismatchTo:(id <HCDescription>)mismatchDescription
 {
     if (![collection conformsToProtocol:@protocol(NSFastEnumeration)])
+    {
+        [[mismatchDescription appendText:@"was non-collection "] appendDescriptionOf:collection];
         return NO;
-    
+    }
+
     if ([collection count] == 0)
+    {
+        [mismatchDescription appendText:@"was empty"];
         return NO;
-    
+    }
+
     for (id item in collection)
+    {
         if (![self.matcher matches:item])
+        {
+            [self describeAllMismatchesInCollection:collection to:mismatchDescription];
             return NO;
+        }
+    }
     return YES;
+}
+
+- (void)describeAllMismatchesInCollection:(id)collection to:(id <HCDescription>)mismatchDescription
+{
+    [mismatchDescription appendText:@"mismatches were: ["];
+    BOOL isPastFirst = NO;
+    for (id item in collection)
+    {
+        if (![self.matcher matches:item])
+        {
+            if (isPastFirst)
+                [mismatchDescription appendText:@", "];
+            [self.matcher describeMismatchOf:item to:mismatchDescription];
+            isPastFirst = YES;
+        }
+    }
+    [mismatchDescription appendText:@"]"];
 }
 
 - (void)describeTo:(id<HCDescription>)description